Add preprocessing directives to allow MPAS-A to be compiled for use as a dycore in CAM.
Note: When compiling the MPAS-A dycore for CAM, MPAS-A physics should be disabled by
setting DO_PHYSICS to an empty string in src/core_atmosphere/Makefile, and -DMPAS_CESM
and -DCAM_DYCORE should be added to the definition of MODEL_FORMULATION in the top-level
Makefile. The build of the stand-alone atmosphere_model executable will fail, but the
pre-processed source should be usable with the CAM build system.