Fix issue in MPAS_io_set_var_indices() when deciding if we can reuse a decomposition.
In MPAS_io_set_var_indices(), before committing to creating a new "decomposition"
with PIO_initdecomp, we first check whether we have already created a decomposition
that is compatible (in terms of type, dimensionality, and degrees of freedom).
This logic assumed that all tasks would agree on whether a matching decomposition
was found.

Under the right circumstances, however, an existing decomposition might match on one
or more tasks, but not on all tasks, causing the latter tasks to hang waiting for
those that exited early from the routine. To avoid this, we don't actually return
at the point where we decide that we can return, but instead set a flag value; at
the end of the logic for checking decompositions, we verify that all tasks agree on
the matching decomposition before returning.

Conflicts:

	src/framework/mpas_io.F