Obtain bit-identical results from atm_adv_coef_compression() on different task counts.
This routine previously sorted the cells involved in the flux computation for each edge,
and since this sorting was done based on local cell index, the order of the cells was
not the same for different MPI task counts, leading to bit-reproducibility issues.

Now, it appears that we can get identical results for any combination of MPI task count
and OpenMP thread count.