Improve numerical accuracy of computation of spherical angles in mpas_atm_advection.F.
This computation involves the semi-perimeter of the triangle formed by the two spherical
arcs. Previously, we computed the edge lengths of this triangle using a dot product, which
proves to be too inaccurate in single-precision (sometimes returning zero length) when
the points defining the triangle are very close together.

This inaccurate arc length computation has been replaced by a more accurate one, namely,
the method used in the arc_length function.