Bug fix to MPAS_io_open() when clobber_file = .false.
When clobber_file = .false., we never want to make changes to existing
files, and we should therefore return as soon as we detect that
a file exists rather than continue to open the file without truncating;
the latter approach can lead to unclosed file descriptors when the program
terminates, leading to bad exit codes.

NB: This commit should be reviewed for inclusion in the 'develop' branch.