Fix race condition in MPAS_io_open() when creating new output files with "no clobber" option.
When creating a new output file with clobber_file=.false. (the default if not specified),
it can occasionally happen that the previously non-existent output file is created by
the call to PIO_createfile() on one or more tasks before other tasks have reached the point
of inquiring whether the output file exists. In such cases, the laggardly tasks detect that
the file exists and return early, causing the MPAS job to hang since not all tasks
participate in subsequent collective I/O calls.

The root cause of this issue appears to be the behavior of some iotype options, which
create output files before all MPI tasks have entered the call to create the file.

Conflicts:
	src/framework/mpas_io.F