Optimize atm_advance_scalars
This commit implements optimizations to the atm_advance_scalars routine,
principally by computing and storing fluxes for the horizontal flux divergence
calculation to avoid computing them twice: once for each of the cells separated
by the edge.

With GNU 4.9.2 compilers, and the 10242 mesh with 41 layers, single-threaded
runtime for this routine was reduced by about 37 percent.