TEMPORARY FIX: for 'xtime' issue with dynamics sub-cycling
At the moment, we are either writing out the wrong time-level to output files,
or the 'xtime' variable represents the time before we write output. This needs
to be looked at more closely, but for now, we can enable restartability by
simply reading the nearest time from the restart file.