Uncomment retrieval of slope from slope_data array in redprm( )
I had commented-out this line a while back because we were getting out-of-bounds
array references due to zero values of the 'slopetyp' index used in the slope_data
array. This was probably due to some fault in our initial conditions file.

Rusty Benson noticed this commented-out line, and after some testing with a recent
intial conditions file, we should revert the commenting of this line.