Declare fields used in GFS physics code as PHYSKIND rather than R8KIND
The change allows us to define PHYSKIND the same as R4KIND when runing
the init_atmosphere core (which uses the monthly time interpolation routine),
but to declare PHYSKIND the same as R8KIND when running the atmosphere model.

For now, we simply define PHYSKIND in the framework mpas_kind_types module
to avoid the need to use a new module in most physics modules.