Temporary fix for cells that are land but have a slope type < 0.5
Currently, our initialization may produce land cells that have a value
for the slope field less than 1 (though <0.5 is the important threshold,
since the code in gbphys adds 0.5 and truncates to an integer the values
in this field); this will cause array index bounds errors in the GFS
physics. As a temporary fix, we set the slope type for such cells to
a value of 9, which is what is assigned to sea-ice points in gbphys.

Ultimately, this should be fixed by using an interpolation that is
consistent with the land-sea mask in mpas_init_atm_static_nggps.F