Add 'avec_benchmark' stream, written at the end of a run only
This commit adds a new mutable stream named 'avec_benchmark' to
the MPAS-Atmosphere core. This streams should have an output_interval
set to "none", since we forcibly write this stream after the last
AVEC timer has finished.

Also included in this commit is a new subroutine, compute_avec_diagnostics,
in the mpas_atm_core_run routine that is called just before the
avec_benchmark stream is written. This routine can be used to compute
any diagnostic fields that we choose to write out for verification
purposes.