Add 'nocnv' option (corrected) in gbphys
This commit adds the option of turning off the SAS convection scheme in the GFS
physics following an e-mail from Jeff Whitaker on 2 March 2016.

The 'nocnv' option is set in mpas_atmphys_nggps_model_parameters.F, and is
passed as an argument to nuopc_phys_init(), where it is stored in
the 'model_parameters' type before it is passed to gbphys( ) by nuopc_phys_run( ).