Add new module mpas_atm_boundaries
This commit adds a new module, mpas_atm_boundaries, in the src/core_atmosphere/dynamics
directory. This module is also symbolically linked to the src/core_init_atmosphere directory.

The initial version of the module contains a single routine, mpas_atm_setup_bdy_masks,
that may be used to set up masks for cells, edges, and vertices on the boundary and in
the relaxation zone of a limited-area domain.