Set up boundary masks in init_atmosphere core
Until we decide on when and where we need to set up boundary masks (e.g.,
does this only need to be done for real-data cases, and only when static
fields are processed?), we simply call the mpas_atm_setup_bdy_masks() routine
from the init_atm_core_init() routine.

Also, add bdyMask{Cell,Edge,Vertex} fields to the default input and output
streams for the init_atmosphere core.