Compute boundary tendencies at the output interval specified for 'lbc_out' stream
The time interval used in the computation of LBC tendencies was previously hard-coded
as 10800 s. Now, the mpas_atm_setup_bdy_tend() and mpas_atm_compute_bdy_tend() routines
check the times at which they are called to automatically determine the interval to
be used in the tendency computation, and this calling interval is determined by the
output interval of the 'lbc_out' stream.

NOTE: At present, the tendency output files will not be correct if the output interval
is not evenly divided by the integration timestep! Fixing this appears to require
modifications to the stream mananger module.