Swap value of 'xtime' to beginning of LBC interval before writing lbc_out stream
The output files written to the 'lbc_out' stream include the 'xtime' field to
identify the beginning time at which the tendencies are valid; however, because
we cannot write the tendencies until some later time, the 'xtime' variable in
these files represented the end of the tendency interval. This causes problems
when reading back in the stream, since the stream manager expects the xtime in
a file to match the time at which the file is read.

This commit implements a simple routine, atm_substitute_xtime, that sets the
xtime variable in the state pool and returns the previously stored value. This
can be used to swap the xtime value before writing the 'lbc_out' stream, then
reset the xtime value after the stream has been written.