Initial implementation of LBC tendency input
This commit implements the basic functionality that will be needed for
periodically reading LBC tendencies and applying those tendencies during
model integration. The key parts of this commit include:

* Define new input stream 'lbc_in'
* Define new namelist option config_apply_lbc_tend in the &lbc record
* Define a new package, lbc_in, attached to the lbc_in stream and tendency
  fields in the Registry
* Add new routine mpas_atm_update_bdy_tend() in the mpas_atm_boundaries
  module that is called immediately after the 'lbc_in' stream is read
* Add code in the atm_core_run routine to read the lbc_in stream and
  call mpas_atm_update_bdy_tend() whenever this stream is read

Because the mpas_atm_update_bdy_tend() routine is called only when
the LBC tendencies are updated by the lbc_in stream, we could in principle
add logic to compute field values based on the previous field value,
the tendency, and the time interval since the tendencies were last updated.