Implement dynamics sub-cycling capability
This commit implements the ability to sub-cycle the dynamics within
a physics timestep. This capability is controlled by a new namelist option,
config_dynamics_subcycles, whose default value is 1. Because this capability
has only been exercised with GFS physics during NGGPS work, this namelist
option does not appear in the default namelist.atmosphere files that are
generated when MPAS-Atmosphere is built.

With this change, 'config_dt' should be interpreted as the physics timestep,
and the timestep for a dynamics subcycle is (config_dt / config_dynamics_subcycles).
Within each dynamics subcycle, dynamics-transport splitting works as before.