Add code to validate runs of the dry baroclinic wave test case
The idea behind this validation code is to compare the min/max/stddev
of the meridional velocity and surface pressure at the end of a simulation
with pre-computed values, and if the relative error is more than some
fixed threshold, to indicate that the run failed.

Currently, a validation routine is called on the last model timestep
before the final output fields are written. This allows us to compute
validation fields and have those fields be included in the final
model output files.

New fields introduced for validation are:
 meridionalMin
 meridionalMax
 meridionalStddev
 psfcMin
 psfcMax
 psfcStddev