If 'make' is run with no target, extract list of available targets automatically
Previously, if 'make' was run with no target, the usage message included
an incomplete list of example targets. To be more helpful, running 'make' with
no options will now automatically extract a list of available build targets
from the Makefile itself.

Available build targets must be identified by adding a comment beginning
with BUILDTARGET on the line of that target; the BUILDTARGET keyword can
be followed by text to describe the applicability or other details of
that target.