Update preprocessor directives to enable both stand-alone builds and CAM builds
The MPAS code now makes use of two preprocessor directives so that one
can easily build a working stand-alone MPAS-Atmosphere model or use
the un-preprocessed source to build CAM.

The directive MPAS_CAM is used in the Registry.xml file to select fields that
need to either exist or to not exist when using the MPAS source for CAM.
If MPAS will be used to build a CAM dycore, -DMPAS_CAM must be added to
the definiton of MODEL_FORMULATION in the top-level Makefile. For stand-alone
MPAS builds, -DMPAS_CAM must not be defined, else certain namelists will
not work and other necessary fields (e.g., land-surface fields) will not be
present.

The directive CAM_DYCORE is defined by the CAM build system to enable code
that is only used when MPAS is running as part of CAM. CAM_DYCORE should
never be defined in the MPAS Makefile, else MPAS will fail to compile
the atmosphere_model executable.

From this branch, it is now possible to pull the *.inc (including those
generated by the Registry in src/core_atmosphere/inc/) and *.F files into
CAM's src/dynamics/mpas/external directory.