Replacing explicit OpenACC with build variable.
I think it's a mistake to embed explicit flags like "-ta=tesla:cc60" in the deep Makefile's. It takes some work to figure out what needs to be changed, so it's too easy to lose track of the individual changes, and get the configuration files into an inconsistent state.
I recommend adding this $(OPENACC) variable to the Makefile instead. I'll be setting it in my top-level build-script so I can manage it as part of my "thin layer".
In the top-level Makefile, you may want to duplicate the "pgi" target as "pgi-openacc" and add this line

OPENACC = -acc -Minfo=accel -ta=tesla:cc60,cuda8.0

I recommend cc60 & cuda8.0 for now because we're working mainly with Pascal's and the cuda9.0 is too new to rely on. On the other hand I won't be using the Makefile setting, I'll be controlling the $(OPENACC) variable from my top-level build script.