Adding $(OPENACC) build variable
(This goes with the analogous change to src/Makefile. I *think* they'll be bundled into the same pull-request).
The OpenACC flags are necessary to compile the files for GPU.
In the top-level Makefile, you may want to duplicate the "pgi" target as "pgi-openacc" and add this line

OPENACC = -acc -Minfo=accel -ta=tesla:cc60,cuda8.0

I recommend cc60 & cuda8.0 for now because we're working mainly with Pascal's and the cuda9.0 is too new to rely on.
I'll be setting the $(OPENACC) variable in my top-level build script ("thin layer").