Simple starting point for running radiation asynchronously
This commit modifies the MPAS-A role functions to define two roles:
0 (GPU work) and 1 (CPU radiation work). The first n/2 MPI tasks perform
role 0, while the second n/2 tasks perform role 1.

In the MPAS-A solver, both roles perform all work at the moment, with
the exception of stream output. Currently, only the role 0 tasks write
output streams to prevent the same file from being simultaneously written
by two sets of MPI tasks.

Note that, if MPAS-A is launched with N MPI tasks, a graph.info.part file
with N/2 partitions is needed.