Move computation of forward-elimation/back-substitution coeffs to acoustic routine
The coefficients for the forward-elimination and back-substitution in
the tri-diagonal solver for the implicit treatment of vertically propagating
acoustic waves were previous computed in the atm_compute_vert_imp_coefs routine
as alpha_tri and gamma_tri; these were then used in the atm_advance_acoustic_step
routine.

Now, we compute the a, b, and c components of the tri-diagonal matrix in
the compute_vert_imp_coefs routine, and use these in the atm_advance_acoustic_step
routine to compute the alpha_tri and gamma_tri arrays, which can now be local
arrays. This should allow for the switch to parallelizable tri-diagonal solvers
that are more performant on GPUs.